Corrected JPQL queries, adapted messaging so that it works. Created a connectionRegistry to keep a list of connections per websocket clients and per player secondary id.

+++ scamlab-server/src/main/java/model/dto/GameDto.java
@@ -16,7 +16,7 @@ public class GameDto {
 
     @RegisterForReflection
     public static record WaitingLobbyStatisticsMessageDto(
-        WSMessageType type, Integer waitingPlayerCount, Integer ongoingGamesCount, Integer maxOngoingGamesCount
+        WSMessageType type, Long waitingPlayerCount, Long ongoingGamesCount, Long maxOngoingGamesCount
     ) {
     }
 }
